Replace XDebug with PCOV and PHPDBG
I just realized that PHP 5.6 onwards included a debugging tool with it
already (PHPDBG), and switching to PCOV for test coverage purposes will
make it a lot faster in case of building CI pipelines in the future.

XDebug took too long to run in CI, because it creates a way larger
overhead.