Update docker image bases
- Remove patch version of composer, now using minor versions
- Move installing mailhog, installing and configuring PCOV and
configuring SSL into 3 different steps. This will make the image a bit
larger in size, but will make it easier to build in the future because
of layer caching.
- Remove the need to generate SSH keys in the production image, as it
will copy the SSL image from dev stage. To run this in production,
ideally there will be a real SSL key to use as a volume.